+++ src/opts/SkBitmapProcState_opts_SSSE3.cpp
@@ -395,7 +395,7 @@ void S32_generic_D32_filter_DX_SSSE3(const SkBitmapProcState& s,
 
     const uint8_t* src_addr =
             static_cast<const uint8_t*>(s.fBitmap->getPixels());
-    const unsigned rb = s.fBitmap->rowBytes();
+    const size_t rb = s.fBitmap->rowBytes();
     const uint32_t XY = *xy++;
     const unsigned y0 = XY >> 14;
     const uint32_t* row0 =
@@ -586,7 +586,7 @@ void S32_generic_D32_filter_DXDY_SSSE3(const SkBitmapProcState& s,
 
     const uint8_t* src_addr =
                         static_cast<const uint8_t*>(s.fBitmap->getPixels());
-    const unsigned rb = s.fBitmap->rowBytes();
+    const size_t rb = s.fBitmap->rowBytes();
 
     // vector constants
     const __m128i mask_dist_select = _mm_set_epi8(12, 12, 12, 12,
